requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
geopy==2.4.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
python-dotenv==1.0.0
supabase==2.3.0
pywebpush==1.14.0
twilio==8.10.3
requests==2.31.0
httpx==0.25.2